+++ src/thread_pool.h
@@ -47,7 +47,7 @@ status_t thread_pool_collect (thread_pool_t *thread_pool, bool cancel);
 status_t thread_pool_cancel (thread_pool_t *thread_pool);
 status_t thread_pool_join (thread_pool_t *thread_pool);
 
-int create_threads (thread_pool_t *thread_pool, int number_of_threads,
+int create_threads (thread_pool_t *thread_pool, long number_of_threads,
                     void *func (void *), void *context, size_t context_size,
                     char *name);
